Refactor book deletion logic into BookDeletionServiceImpl
- Moved book deletion and response handling logic from BookController to BookDeletionServiceImpl.
- BookDeletionService now returns a ResponseEntity directly, encapsulating the response creation.
- Controller delegates the deletion operation to the service layer, simplifying its responsibilities.
- Updated BookController to use the service layer for deletion and response handling.
