+++ library-sql/src/main/java/pl/library/library/Main.java
@@ -1,4 +1,4 @@
-package pl.library.library.services;
+package pl.library.library;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
