+++ library-sql/src/main/java/pl/library/library/services/BookDeletionService.java
@@ -1,0 +1,8 @@
+package pl.library.library.services;
+
+
+import org.springframework.http.ResponseEntity;
+
+public interface BookDeletionService {
+    ResponseEntity<String> deleteBook(Long id);
+}
